backend/pyproject.toml
[project]
name = "build-risk-backend"
version = "0.1.0"
description = "FastAPI backend service for the BuildGuard DevSecOps risk dashboard."
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [{ name = "BuildGuard Team" }]
dependencies = [
  "fastapi>=0.110,<0.113",
  "uvicorn[standard]>=0.24,<0.28",
  "pydantic>=2.4,<2.7",
  "pydantic-settings>=2.1,<2.4",
  "pymongo>=4.6,<5.0",
  "httpx>=0.25,<0.28",
  "PyGithub>=2.1,<2.3",
  "python-dotenv>=1.0,<2.0",
  "python-jose[cryptography]>=3.3,<3.4",
  "passlib[bcrypt]>=1.7,<1.8",
  "python-multipart>=0.0.6,<0.0.9",
  "pandas>=2.1,<2.3",
  "numpy>=1.26,<1.28",
  "celery>=5.3,<5.4",
  "tenacity>=8.2,<9.0",
  "GitPython>=3.1,<3.2",
  "python-dateutil>=2.8,<3.0",
  "redis>=5.0,<6.0",
  "pyarrow>=22.0.0",
  "sf-hamilton>=1.62",
]

[project.optional-dependencies]
ml = [
  "torch>=2.1,<2.2",
  "torchvision>=0.16,<0.17"
]
dev = [
  "pytest>=7.4,<8.2",
  "pytest-asyncio>=0.21,<0.24",
  "httpx>=0.25,<0.28",
  "black>=23.12,<24.11",
  "ruff>=0.5.0,<0.6.0",
  "mypy>=1.7,<1.10"
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
  "pytest>=7.4,<8.2",
  "pytest-asyncio>=0.21,<0.24",
  "httpx>=0.25,<0.28",
  "black>=23.12,<24.11",
  "ruff>=0.5.0,<0.6.0",
  "mypy>=1.7,<1.10"
]

[tool.setuptools]
packages = ["app"]
